requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
faiss-cpu==1.9.0.post1
numpy==2.1.3
openai==1.58.1
httpx==0.28.1
